requirements.txt
nltk
PyPDF2
pandas
gspread
google-auth
sentence-transformers
torch
scikit-learn
google-genai
